EsLint and Prettier

+++ server/schemas/resolvers.js
@@ -1,7 +1,8 @@
 const { Event, User, Comment } = require('../models');
-// auth 
+// auth
 
 const resolvers = {
+
   Query: {
     getEventData: async (parent, { eventInput }) => {
       return Event.findOne({_id: eventInput._id});
@@ -26,17 +27,17 @@ const resolvers = {
     //     if (!user) {
     //         throw console.log("error");
     //     }
-    //     // need auth 
+    //     // need auth
     // },
     deleteUser: async (parent, { userID }, context) => {
-      if (context.user)
-        return User.findOneAndDelete({ _id: userID })
+      if (context.user) return User.findOneAndDelete({ _id: userID });
 
-      throw new Error("Something has gone wrong!");
+      throw new Error('Something has gone wrong!');
     },
 
     addEvent: async (parent, { eventInput }, context) => {
-      if (context.user) { // needs testing
+      if (context.user) {
+        // needs testing
         const event = await Event.create(eventInput);
 
         await User.findByIdAndUpdate(context.user._id, {
@@ -46,23 +47,24 @@ const resolvers = {
         return event;
       }
 
-      throw new Error("Something has gone wrong!");
-
+      throw new Error('Something has gone wrong!');
     },
 
     updateEvent: async (parent, { eventInput }, context) => {
       if (context.user) {
-        const event = await Event.findOneAndUpdate({ _id: eventInput._id },
+        const event = await Event.findOneAndUpdate(
+          { _id: eventInput._id },
           {
-            $set: eventInput
+            $set: eventInput,
           },
           {
             runValidators: true,
             new: true,
           }
-        )
+        );
         return event;
       }
+
       throw new Error("Not logged in");
     },
 
@@ -135,4 +137,4 @@ const resolvers = {
   }
 }
 
-module.exports = resolvers
+module.exports = resolvers;
